Refactor: move useUpdateComments hook to sendCommentForm
